cpufreq: schedutil: Cast TICK_USEC to u64
This fixes the following warning:

../kernel/sched/cpufreq_schedutil.c:402:11: warning: comparison of distinct pointer types ('typeof (rq->curr->se.slice / 1000) *' (aka 'unsigned long long *') and 'typeof (((1000000L + 300 / 2) / 300)) *' (aka 'long *')) [-Wcompare-distinct-pointer-types]
  402 |                 delay = min(rq->curr->se.slice/1000, TICK_USEC);
      |                         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../include/linux/kernel.h:908:19: note: expanded from macro 'min'
  908 | #define min(x, y)       __careful_cmp(x, y, <)
      |                         ^~~~~~~~~~~~~~~~~~~~~~
../include/linux/kernel.h:899:24: note: expanded from macro '__careful_cmp'
  899 |         __builtin_choose_expr(__safe_cmp(x, y), \
      |                               ^~~~~~~~~~~~~~~~
../include/linux/kernel.h:889:4: note: expanded from macro '__safe_cmp'
  889 |                 (__typecheck(x, y) && __no_side_effects(x, y))
      |                  ^~~~~~~~~~~~~~~~~
../include/linux/kernel.h:875:29: note: expanded from macro '__typecheck'
  875 |                 (!!(sizeof((typeof(x) *)1 == (typeof(y) *)1)))
      |                            ~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~~

+++ kernel/sched/cpufreq_schedutil.c
@@ -300,7 +300,7 @@ static inline unsigned long sugov_apply_dvfs_headroom(unsigned long util, int cp
 	 * switch or TICK?
 	 */
 	if (rq->cfs.h_nr_running > 1)
-		delay = min(rq->curr->se.slice/1000, TICK_USEC);
+		delay = min(rq->curr->se.slice/1000, (u64)TICK_USEC);
 	else
 		delay = TICK_USEC;
 	delay = max(delay, per_cpu(dvfs_update_delay, cpu));
